working with a continuous substrate potential and toying around with icc

+++ substrate.h
@@ -2,5 +2,5 @@
  * @file substrate.h
  * @brief Everything dealing with the substrate
  */
-int external_energy(vector2d);
+double external_energy(vector2d);
 void init_substrate(void);
